cropped and renamed static images, finished styling contact
Cropped the images to make styling a lot easier, added Contact to default.vue because it seems to be found in every frame in the panel.
If you resize the window to a smaller size, the bunchofleaves is set to not show, because if it did you'd have to transform and style it in so many confusing ways that'd it be easier to just include another image that's just rotated, so I hid it. If we use another image that's just a rotated bunchofleaves, then it can be easily implemented.